Debian Image Builder
Host: Ubuntu Focal
Updated u-boot to v2021.01
Updated u-boot functions

Ubuntu Focal offers more options in the GCC department and will
enable uboot and kernels to be built specific for the release
of Debian that is chosen.

NOTE:
Don't use this branch! As I've not updated any kernel functions as
of yet.