Option to force KBUILD_BUILD_VERSION
Unfortunately this isn't very well documented, but essentially we can change the BUILD VERSION of the kernel. Anything placed in the input box will be appended to the end of the kernel version of the Linux deb package. By default this would simply be the number one and when using dmesg would be seen as # 1.