First Boot: `Vendor VS Mainline`
The script will no longer check for vfat on mmcblk2. Why vendors are still using vfat partitions for boot is beyond me on such boards, but apparently it appears to be creating issues on first boot if a Vendor IMG is installed on MMC and then one of our IMGS it booted from SD. Makes me think maybe the MMC should be hit with a DD to correct any further trouble so the boot process is fresh.