Radxa Zero: `tweaks`
I'm not 100% sure why, but for whatever reason when over clocking in Debian, it doesn't always kick in until the governor script is run. I have a feeling this has something to do with the sleep added to the hdmi audio bit. That bit may need to become its own service?