patches: uboot: v2023.07.02: bananapicm4: `boot target nvme`
HACK: BOOT ORDER: NVMe SDCARD eMMC.

NOTES:
In my testing there has been no false starts or hangs up. Meaning
the boot process has been stable.

The downside to this in my opinion is that if there is an OS on
the NVMe it will always take boot priority. The drive would need
to be pulled in order for SD eMMC boot to kick in.

Tested-on: Waveshare CM4-IO-BASE-B
Signed-off-by: Patrick Yavitz <pyavitz@xxxxx.com>