U-Boot and Linux: `packaging`
Kernel packaging will now be handled in two different ways.
If Linux patching and defconfigs are shared, than a common
packaging name will be used. Currently there are only two
platforms this applies to; Amlogic "meson" and Allwinner
"sunxi". The goal is to eventually have as many platforms
share common packaging and kernels.

The benefit of common packaging; One can build a kernel for
the Odroid C4 and install it on the Banana Pi CM4 and have
no loss in performance or stability.

If there is no shared patching than the kernel will continue
to use the packaging name based on the SoC series.

U-Boot packaging names have changed and will now include the
$BOARD variable in the name of the deb built.

NOTE: Units effected by this change will need to purge the
old linux and u-boot pkgs before installing the new.

Signed-off-by: Patrick Yavitz <pyavitz@xxxxx.com>