`Debian Image Builder`
When dealing with git, the builder will no longer download
archives and instead git clone the repo. This is a bit more
time consuming, but provides more options and the ability to
download sources from locations other than github.

This does come with a few caveats. The ability to build vendor
kernels on a whim had to be sacrificed. But in time the feature
should be able to be added back.

ROCK5B is now pulling Collabora by default.

Signed-off-by: Patrick Yavitz <pyavitz@xxxxx.com>