BananaPi F3: BSP: `defconfig and patching`
Introduced a new linux function: $FAMILY_bsp

In this case I'm using it to change the PATCHDIR and DEFCONFIG
vars when linux VERSION=6.1.15 is used.

Currently the board file is set to pull 6.1.y by default. So
we need to force_version=false in order to download the kernel.

Example:
make kernel board=bananapif3 version=6.1.15 force_version=false

Signed-off-by: Patrick Yavitz <pyavitz@xxxxx.com>