Allwinner H616 has a dedicated power domain for its Mali G31.
Currently after probe, the GPU is put in runtime suspend which
disable the power domain.
On first usage of GPU, the power domain enable hangs the system.

This series adds the necessary calls to enable the clocks and
deasserting the reset line after the power domain enabling and
asserting the reset line and disabling the clocks prior to the
power domain disabling.

This allows to use the Mali GPU on all Allwinner H616
boards and devices.

https://lore.kernel.org/linux-sunxi/20250312232319.25712-1-simons.philippe@gmail.com/T/#t

Signed-off-by: Patrick Yavitz <pyavitz@gmail.com>