Rockchip: `Update u-boot build process for RK358X inclusion`
Added a generic rock5b board conf as an example for future adds.
There are currently no patches in place for u-boot or the kernel.

U-Boot: set to `v2025.07`
Kernel: set to `stable`

The current linux defconfig may or may not need to be updated
to work with RK358X.

Signed-off-by: Patrick Yavitz <pyavitz@gmail.com>